Make MirrorMaker optional in CLI
